task edited

+++ circuleArea.py
@@ -1,5 +1,5 @@
 pi = 3.14158
 
-radius = int(input("Enter the radius of the circle: "))
+radius = float(input("Enter the radius of the circle: "))
 
 print("The area of the circule is: ",str(pi*(radius*radius)) )
